scripts/requirements.txt
kafka-python==2.2.15
redis==6.4.0
